requirements.txt
# Skej - Voice Assistant Project
# Core backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pandas>=1.5.0
python-multipart
jinja2
python-dotenv

# WebRTC & Voice Pipeline Dependencies
aiortc>=1.9.0
aiohttp>=3.9.0
sounddevice>=0.4.6
av>=11.0.0
websockets>=11.0.0
numpy>=1.24.0
soundfile>=0.12.1
openai>=1.45.0

# Optional: For local voice processing (uncomment if needed)
# openai-whisper>=20231117
# torch>=2.0.0
# torchaudio>=2.0.0

# Development tools
pytest>=7.0.0
black
flake8